fixed bundling issue and bundled with webpack
