+++ index.html
@@ -208,19 +208,18 @@
 
     </div>
 
-
-
     <div class="overlay"></div>
 
+    <script type="module" src="dist/bundle.js"></script>
+</body>
+
+</html>
 
 
 
-    <script type="module" src="dist/constants.js"></script>
+<!-- <script type="module" src="dist/constants.js"></script>
     <script type="module" src="dist/type.js"></script>
     <script type="module" src="dist/script.js"></script>
     <script type="module" src="dist/filterAndSearchFun.js"></script>
     <script type="module" src="dist/SortFun.js"></script>
-    <script type="module" src="dist/tableActionButton.js"></script>
-</body>
-
-</html>
+    <script type="module" src="dist/tableActionButton.js"></script> -->
