+++ webpack.config.js
@@ -1,6 +1,5 @@
 //webpack.config.js
 const path = require('path');
-const webpack = require('webpack')
 module.exports = {
     mode: "development",
     devtool: "inline-source-map",
